Fixed enum values so the following enum
typedef enum
{
    Z_A,
    Z_B = 0x2B,
    Z_C = 'j',
    Z_D,
} Zebra;

gets interpreted as
[{'name': 'Z_A', 'value': 0},
 {'name': 'Z_B', 'value': 43},
 {'name': 'Z_C', 'value': 106},
 {'name': 'Z_D', 'value': 107}]

Previously the auto resolver would not notice that Z_B and Z_C would affect the auto incrementer.

Additionally, values are now converted to an int as C++ actually sees them.

If you want to maintain the original format of the enum, pass in the keyword arg enumMaintianValueFormat=True to the CppHeader constructor

Also converted tabs to 4 spaces.

-Jash